Optionally cap number of observations used to compute center of mass. (#8)
This attempts to mitigate some of the quadratic complexity when dealing with
datasets containing many batches. The idea is to cap the size of the reference
dataset when computing the center of mass, ensuring that it is agnostic to the
increase in the size of the reference at later merge steps. We achieve this
capping by just taking every n-th cell, effectively downsampling the dataset.