Refactor to a free-standing compute() function and an Options class. (#11)
This is more intuitive to understand compared to the previous class methods.
We move the optional ordering argument into our new Options class, and we add a
Builder member to allow users to specify their own KNN algorithm.

The underlying functionality is more or less the same. The only changes are:

- We do not cap the target's NN search when computing the correction vectors;
  this is because we need NNs for each target cell to correct it. The cap is
  now only applied to the reference, as NNs are not required for each cell. 
- We switch from end to length in custom parallelization, for consistency
  with other libraries in the libscran organization.

Fleshed out the C++ tests for almost every part of the codebase, now with all
warnings enabled. It also builds the tests for OpenMP (if available) and custom
parallelization by default, allowing us to streamline the CI actions. Added
more tests for the reference R implementation to check options and ordering.
Also added a minor patch to the real data example to get it to work for the
time being, at least until we replace scran.chan with scrapper.

Added install commands to the CMakeLists, with an associated CI action for
testing installation. Removed the unnecessary direct dependency on aarand. 

Also updated the README for the organization move. Fleshed out the quick start
section and added some theoretical details about the mass cap.