Overhauled integer types to use Index_ instead of std::size_t for observations. (#13)
This is more appropriate and ensures that we are consistent with the NN output.
We need to cast back to size_t to avoid overflow when computing array offsets,
which shouldn't be too expensive; any perf losses are probably outweighed by
faster increments on an Index_ counter so it all washes out in the end.

So, size_t's are now limited to the 2-D array offsets and the batch indices
(the latter of which are typedef'd as BatchIndex for clarity). We make sure to
include <cstddef> and add a namespace (i.e., std::size_t) for compliance.

As mass_cap parameter is defined in terms of the number of observations, we
also turn this into an Index_. However, doing so means that the default
mass_cap = -1 is not safe; for signed integers, it will not overflow to give a
very large cap that can be effectively ignored. Instead, we default to mass_cap
= 0 and explicitly ignore the cap if the provided value is <= 0.

Finally, we remove the default int and double types in the Options template
parameters. Users should just take some time to read the docs and specify it
themselves to make sure that they get the expected input/output.