Yet more appropriate choices of integer types in robust_average.
We template the vector counter (and rename the variables) to indicate that
the caller is responsible for figuring out the best type.

We also switch from resize() to insert() when building the data, in order to
give the library a chance to throw bad_alloc if it overflows the size_type.