More stringent filtering of outliers during center-of-mass calculations. (#15)
The new approach iteratively adds more points to the center of mass,
calculating the running mean and variance; for each new point (sorted by
distance from the MNN-involved cell), we only add it to the center of mass if
it is less than some multiple of standard deviations from the mean. This allows
us to filter out points from other subpopulations. 

By comparison, the old approach (median + MADs on the distance distribution)
doesn't work too well at high dimensions as all distances end up being similar.
The use of a median also assumes that most neighbors of MNN-involved cells
belong to the same subpopulation, which effectively requires that the majority
of cells belong in shared subpopulations between batches.