Refactor algorithm to search for MNNs against all other batches at once. (#17)
Specifically, every observation in the target batch gets corrected to one of
the remaining batches, based on the identity of the other observation in the
closest MNN pair. This reduces the risk of MNN pairs forming between the wrong
subpopulations because we didn't choose the right reference or merge order. 

We also refactored the center of mass calculations to just use the centroid of
the neighbors of each MNN-involved observation. This reduces the risk of adding
too many observations into the center of mass, particularly problematic in
cases where the observations are continuously distributed but the continuum is
not completely present in all batches. Users can tune this by performing more
steps in the recursive neighbor search if they want to avoid kissing effects.

We updated the R test scripts to include some examples of merging >2 batches.